+++ sydtest-webdriver-yesod/default.nix
@@ -1,15 +1,16 @@
 { mkDerivation, base, bytestring, http-client, http-types, lib, mtl
 , network-uri, path, path-io, sydtest, sydtest-discover
-, sydtest-wai, sydtest-webdriver, sydtest-yesod, text, webdriver
-, yesod
+, sydtest-wai, sydtest-webdriver, sydtest-yesod, text, wai
+, webdriver, yesod
 }:
 mkDerivation {
   pname = "sydtest-webdriver-yesod";
-  version = "0.0.0.1";
+  version = "0.0.1.0";
   src = ./.;
   libraryHaskellDepends = [
     base bytestring http-client http-types mtl network-uri sydtest
-    sydtest-wai sydtest-webdriver sydtest-yesod text webdriver yesod
+    sydtest-wai sydtest-webdriver sydtest-yesod text wai webdriver
+    yesod
   ];
   testHaskellDepends = [
     base path path-io sydtest sydtest-webdriver yesod