+++ sydtest-yesod/src/Test/Syd/Yesod/Def.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE RankNTypes #-}
@@ -13,17 +14,24 @@ module Test.Syd.Yesod.Def
     yesodSpecWithSiteSupplier,
     yesodSpecWithSiteSupplierWith,
     yesodSpecWithSiteSetupFunc,
+    yesodSpecWithSiteSetupFuncWithMiddlewares,
     yesodSpecWithSiteSetupFunc',
+    yesodSpecWithSiteSetupFuncWithMiddlewares',
     yesodClientSetupFunc,
+    yesodClientSetupFuncWithMiddlewares,
     YesodSpec,
     yit,
     ydescribe,
   )
 where
 
+#if !MIN_VERSION_base(4,20,0)
+import Data.List (foldl')
+#endif
 import GHC.Stack (HasCallStack)
 import Network.HTTP.Client as HTTP
 import Network.URI
+import Network.Wai (Middleware)
 import Test.Syd
 import Test.Syd.Wai
 import Test.Syd.Yesod.Client
@@ -148,6 +156,15 @@ yesodSpecWithSiteSetupFunc ::
   TestDef (HTTP.Manager ': outers) ()
 yesodSpecWithSiteSetupFunc setupFunc = yesodSpecWithSiteSetupFunc' $ \man () -> setupFunc man
 
+-- `yesodSpecWithSiteSetupFunc` with middlewares
+yesodSpecWithSiteSetupFuncWithMiddlewares ::
+  (YesodDispatch site) =>
+  [Middleware] ->
+  (HTTP.Manager -> SetupFunc site) ->
+  TestDef (HTTP.Manager ': outers) (YesodClient site) ->
+  TestDef (HTTP.Manager ': outers) ()
+yesodSpecWithSiteSetupFuncWithMiddlewares middlewares setupFunc = yesodSpecWithSiteSetupFuncWithMiddlewares' middlewares $ \man () -> setupFunc man
+
 -- | Using a function that supplies a 'site', using a 'SetupFunc'.
 --
 -- This function assumed that you've already set up the 'HTTP.Manager' beforehand using something like 'managerSpec'.
@@ -160,9 +177,23 @@ yesodSpecWithSiteSetupFunc' setupFunc = setupAroundWith' $ \man inner -> do
   site <- setupFunc man inner
   yesodClientSetupFunc man site
 
+yesodSpecWithSiteSetupFuncWithMiddlewares' ::
+  (YesodDispatch site) =>
+  [Middleware] ->
+  (HTTP.Manager -> inner -> SetupFunc site) ->
+  TestDef (HTTP.Manager ': outers) (YesodClient site) ->
+  TestDef (HTTP.Manager ': outers) inner
+yesodSpecWithSiteSetupFuncWithMiddlewares' middlewares setupFunc = setupAroundWith' $ \man inner -> do
+  site <- setupFunc man inner
+  yesodClientSetupFuncWithMiddlewares middlewares man site
+
 yesodClientSetupFunc :: (YesodDispatch site) => HTTP.Manager -> site -> SetupFunc (YesodClient site)
-yesodClientSetupFunc man site = do
-  application <- liftIO $ Yesod.toWaiAppPlain site
+yesodClientSetupFunc = yesodClientSetupFuncWithMiddlewares []
+
+yesodClientSetupFuncWithMiddlewares :: (YesodDispatch site) => [Middleware] -> HTTP.Manager -> site -> SetupFunc (YesodClient site)
+yesodClientSetupFuncWithMiddlewares middlewares man site = do
+  application' <- liftIO $ Yesod.toWaiAppPlain site
+  let application = foldl' (\app middleware -> middleware app) application' middlewares
   p <- applicationSetupFunc application
   let client =
         YesodClient