+++ sydtest-webdriver-yesod/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.0.1.0] - 2025-12-29
+
+### Added
+
+* Support for adding Yesod Middlewares to your Yesod application within testing
+
 ## [0.0.0.1] - 2023-02-18
 
 * `currentRouteShouldBe` for route assertions